Move function to run external commands into separate file
